+++ packages/components/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v11.0.0
+## Make "body" of Option component more generic and flexible for use
+Just renamed `description` prop to `content` and remove `<p>` wrapping.
+
 # v10.6.1
 ## Bug fix: flex child props fix.